Rename usb connection
